Add json output, move headers
