+++ cpp/CMakeLists.txt
@@ -8,40 +8,45 @@ add_library(mlt-cpp STATIC)
 set_property(TARGET mlt-cpp PROPERTY CXX_STANDARD 20)
 target_compile_definitions(mlt-cpp PUBLIC PROTOZERO_USE_VIEW=std::string_view)
 
+option(MLT_WITH_JSON "Include JSON support" ON)
+
 target_include_directories(mlt-cpp
     PUBLIC ${PROJECT_SOURCE_DIR}/include
     PRIVATE ${PROJECT_SOURCE_DIR}/src
 )
 
 list(APPEND INCLUDE_FILES
-    ${PROJECT_SOURCE_DIR}/include/common.hpp
-    ${PROJECT_SOURCE_DIR}/include/decoder.hpp
-    ${PROJECT_SOURCE_DIR}/include/feature.hpp
-    ${PROJECT_SOURCE_DIR}/include/geometry.hpp
-    ${PROJECT_SOURCE_DIR}/include/layer.hpp
-    ${PROJECT_SOURCE_DIR}/include/tile.hpp
-    ${PROJECT_SOURCE_DIR}/include/metadata/stream.hpp
-    ${PROJECT_SOURCE_DIR}/include/metadata/tileset.hpp
-    ${PROJECT_SOURCE_DIR}/include/util/buffer_stream.hpp
+    ${PROJECT_SOURCE_DIR}/include/mlt/common.hpp
+    ${PROJECT_SOURCE_DIR}/include/mlt/decoder.hpp
+    ${PROJECT_SOURCE_DIR}/include/mlt/feature.hpp
+    ${PROJECT_SOURCE_DIR}/include/mlt/geometry.hpp
+    ${PROJECT_SOURCE_DIR}/include/mlt/geojson.hpp
+    ${PROJECT_SOURCE_DIR}/include/mlt/layer.hpp
+    ${PROJECT_SOURCE_DIR}/include/mlt/polyfill.hpp
+    ${PROJECT_SOURCE_DIR}/include/mlt/projection.hpp
+    ${PROJECT_SOURCE_DIR}/include/mlt/tile.hpp
+    ${PROJECT_SOURCE_DIR}/include/mlt/metadata/stream.hpp
+    ${PROJECT_SOURCE_DIR}/include/mlt/metadata/tileset.hpp
+    ${PROJECT_SOURCE_DIR}/include/mlt/util/buffer_stream.hpp
 )
 
 list(APPEND SRC_FILES
-    ${PROJECT_SOURCE_DIR}/src/decoder.cpp
-    ${PROJECT_SOURCE_DIR}/src/decode/geometry.hpp
-    ${PROJECT_SOURCE_DIR}/src/decode/int.cpp
-    ${PROJECT_SOURCE_DIR}/src/decode/int.hpp
-    ${PROJECT_SOURCE_DIR}/src/decode/property.hpp
-    ${PROJECT_SOURCE_DIR}/src/decode/string.hpp
-    ${PROJECT_SOURCE_DIR}/src/feature.cpp
-    ${PROJECT_SOURCE_DIR}/src/metadata/stream.cpp
-    ${PROJECT_SOURCE_DIR}/src/metadata/tileset.cpp
-    ${PROJECT_SOURCE_DIR}/src/util/morton.hpp
-    ${PROJECT_SOURCE_DIR}/src/util/raw.hpp
-    ${PROJECT_SOURCE_DIR}/src/util/rle.cpp
-    ${PROJECT_SOURCE_DIR}/src/util/rle.hpp
-    ${PROJECT_SOURCE_DIR}/src/util/varint.hpp
-    ${PROJECT_SOURCE_DIR}/src/util/vectorized.hpp
-    ${PROJECT_SOURCE_DIR}/src/util/zigzag.hpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/decoder.cpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/decode/geometry.hpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/decode/int.cpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/decode/int.hpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/decode/property.hpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/decode/string.hpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/feature.cpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/metadata/stream.cpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/metadata/tileset.cpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/util/morton.hpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/util/raw.hpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/util/rle.cpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/util/rle.hpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/util/varint.hpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/util/vectorized.hpp
+    ${PROJECT_SOURCE_DIR}/src/mlt/util/zigzag.hpp
 )
 
 target_sources(mlt-cpp PRIVATE
@@ -69,5 +74,18 @@ FetchContent_MakeAvailable(fastpfor)
 target_include_directories(mlt-cpp PRIVATE "${fastpfor_SOURCE_DIR}/headers")
 target_compile_options(FastPFOR PRIVATE "-Wno-cast-align")
 
-add_subdirectory(${PROJECT_SOURCE_DIR}/test)
+# json
+if(MLT_WITH_JSON)
+    message(STATUS "[MLT] Including GeoJSON support")
+    fetchcontent_declare(json
+        GIT_REPOSITORY https://github.com/nlohmann/json.git
+        GIT_TAG v3.11.3
+    )
+    FetchContent_MakeAvailable(json)
+    target_include_directories(mlt-cpp PUBLIC "${json_SOURCE_DIR}/include")
+    target_compile_definitions(mlt-cpp PUBLIC MLT_WITH_JSON=1)
+else()
+    message(STATUS "[MLT] No GeoJSON support")
+endif(MLT_WITH_JSON)
 
+add_subdirectory(${PROJECT_SOURCE_DIR}/test)
