--- cpp/include/common.hpp
@@ -1,51 +1,0 @@
-#pragma once
-
-#include <cstdint>
-#include <iostream>
-#include <string_view>
-#include <type_traits>
-#include <utility>
-
-namespace mlt {
-
-using DataView = std::string_view;
-
-using count_t = std::uint32_t;
-
-template <typename T, std::size_t N>
-constexpr std::size_t countof(T (&)[N]) {
-    return N;
-}
-
-/// `std::underlying_type` that doesn't fail when given a simple type
-template <typename T, bool = std::is_enum_v<T>>
-struct underlying_type {
-    using type = T;
-};
-template <typename T>
-struct underlying_type<T, true> : ::std::underlying_type<T> {};
-template <class T>
-using underlying_type_t = typename underlying_type<T>::type;
-
-#define MLT_LOG_WARN(s) std::cerr << s << '\n'
-
-} // namespace mlt
-
-namespace std {
-#if !__has_cpp_attribute(__cpp_lib_to_underlying)
-template <typename E>
-constexpr auto to_underlying(E e) noexcept {
-    return static_cast<std::underlying_type_t<E>>(e);
-}
-#endif
-
-#if !__has_cpp_attribute(__cpp_lib_byteswap)
-template <std::integral T>
-constexpr T byteswap(T value) noexcept {
-    static_assert(std::has_unique_object_representations_v<T>, "T may not have padding bits");
-    auto value_representation = std::bit_cast<std::array<std::byte, sizeof(T)>>(value);
-    std::ranges::reverse(value_representation);
-    return std::bit_cast<T>(value_representation);
-}
-#endif
-} // namespace std
