--- cpp/src/decode/int.cpp
@@ -1,3 +1,0 @@
-#include <decode/int.hpp>
-
-namespace mlt::decoder {} // namespace mlt::decoder
