+++ cpp/src/mlt/decode/int.cpp
@@ -1,0 +1,3 @@
+#include <mlt/decode/int.hpp>
+
+namespace mlt::decoder {} // namespace mlt::decoder
