+++ cpp/test/decode.cpp
@@ -1,15 +1,21 @@
 #include <gtest/gtest.h>
 
-#include <decoder.hpp>
-#include <metadata/tileset.hpp>
+#include <mlt/decoder.hpp>
+#include <mlt/metadata/tileset.hpp>
+#include <mlt/projection.hpp>
 
 #include <iostream>
 #include <filesystem>
 #include <fstream>
 #include <regex>
 #include <string>
+
 using namespace std::string_literals;
 using namespace std::string_view_literals;
+
+#if MLT_WITH_JSON
+#include <mlt/geojson.hpp>
+#endif
 
 namespace {
 std::vector<std::filesystem::path> findFiles(const std::filesystem::path& base, const std::regex& pattern) {
@@ -61,7 +67,54 @@ std::optional<mlt::MapLibreTile> loadTile(const std::string& path) {
     }
 
     mlt::decoder::Decoder decoder;
-    return decoder.decode({buffer.data(), buffer.size()}, metadata);
+    auto tile = decoder.decode({buffer.data(), buffer.size()}, metadata);
+
+#if MLT_WITH_JSON
+    auto jsonBuffer = loadFile(path + ".geojson");
+    if (!jsonBuffer.empty()) {
+        using json = nlohmann::json;
+        const json::parser_callback_t callback = nullptr;
+        const json expectedJSON = json::parse(
+            jsonBuffer, nullptr, /*allow_exceptions=*/false, /*ignore_comments=*/true);
+        const auto actualJSON = mlt::GeoJSON::toGeoJSON(tile, {3, 5, 7});
+        auto diffJSON = json::diff(expectedJSON, actualJSON);
+
+        // Eliminate differences due to extremely small floating-point value changes
+        for (auto i = diffJSON.begin(); i != diffJSON.end();) {
+            auto& node = *i;
+            assert(node.is_object());
+            if (node["op"] == "replace" && node["value"].is_number()) {
+                const auto& path = node["path"];
+                if (path.is_string()) {
+                    const auto ptr = json::json_pointer{path};
+                    const auto& expectedNode = expectedJSON.at(ptr);
+                    const auto& actualNode = actualJSON.at(ptr);
+                    if (expectedNode.is_number_float() && actualNode.is_number_float()) {
+                        const double expectedValue = expectedNode;
+                        const double actualValue = actualNode;
+                        const double error = std::fabs(expectedValue - actualValue) / actualValue;
+                        if (error < 1.0e-15) {
+                            i = diffJSON.erase(i);
+                            continue;
+                        } else {
+                            node["previous"] = expectedNode;
+                            node["error"] = error;
+                        }
+                    }
+                }
+            }
+            ++i;
+        }
+
+        std::cout << path
+                  << "\n"
+                  //<< "expected=" << expectedJSON.dump(2, ' ') << "\n"
+                  //<< "actual=" << actualJSON.dump(2, ' ') << "\n"
+                  << " diff=" << diffJSON.dump(2, ' ') << "\n";
+    }
+#endif // MLT_WITH_JSON
+
+    return tile;
 }
 
 } // namespace
@@ -72,14 +125,13 @@ TEST(Decode, SimplePointBoolean) {
 
     const auto* mltLayer = tile->getLayer("layer");
     EXPECT_TRUE(mltLayer);
-    EXPECT_EQ(mltLayer->getVersion(), 1); // TODO: doesn't match JS version
     EXPECT_EQ(mltLayer->getName(), "layer");
-    EXPECT_EQ(mltLayer->getTileExtent(), 4096);
+    EXPECT_EQ(mltLayer->getVersion(), 1); // TODO: doesn't match JS version
+    EXPECT_EQ(mltLayer->getExtent(), 4096);
     EXPECT_EQ(mltLayer->getFeatures().size(), 1);
 
     const auto& feature = mltLayer->getFeatures().front();
     EXPECT_EQ(feature.getID(), 1);
-    EXPECT_EQ(feature.getExtent(), 4096);
     // TODO: geometry
     // TODO: properties
     // TODO: GeoJSON
